Cn9 create employee list (#17)
* CN9 added controller to retrieve paginated employees

* CN9 added employee page to front end and updated data seeding

* CN9 added confirmation popup for deletion

* CN9 added UserDTO with location address

* CN9 added display of location in employee table

* CN9 made it possible to delete employee

* CN9 resolved dependency issues

* CN9 added mapping to seeder methods and added extension to handle null config value refferences

+++ CafeNet/Data/Repositories/UserRepository.cs
@@ -1,4 +1,5 @@
 ﻿using CafeNet.Data.Database;
+using CafeNet.Data.Enums;
 using CafeNet.Data.Models;
 using Microsoft.EntityFrameworkCore;
 
@@ -16,9 +17,14 @@ namespace CafeNet.Data.Repositories
             _context.Users.Add(user);
         }
 
-        public void Delete(User user)
+        public void DeleteById(long id)
         {
-            _context.Users.Remove(user);
+            var user = _context.Users.FirstOrDefault(u => u.Id == id);
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+                _context.SaveChanges();
+            }
         }
         public async Task<User> GetByIdAsync(long id)
         {
@@ -40,6 +46,23 @@ namespace CafeNet.Data.Repositories
             await _context.SaveChangesAsync();
             return user;
         }
+
+        public async Task<IEnumerable<User>> GetByRolesPagedAsync(IEnumerable<UserRoles> roles, int pageNumber, int pageSize)
+        {
+            return await _context.Users
+                                .Include(u => u.Location)
+                                .Where(u => roles.Contains(u.Role))
+                                .Skip((pageNumber - 1) * pageSize)
+                                .Take(pageSize)
+                                .ToListAsync();
+        }
+
+        public async Task<int> CountByRolesAsync(IEnumerable<UserRoles> roles)
+        {
+            return await _context.Users.CountAsync(u => roles.Contains(u.Role));
+        }
+
+
         public async Task<bool> UsernameExistsAsync(string username)
         {
             return await _context.Users.AnyAsync(u => u.Username == username);
